+++ app/index.js
@@ -16,13 +16,11 @@ var ProtractorGenerator = module.exports = function ProtractorGenerator(args, op
     });
 
     console.log('Done! Now follow these steps:'.bold);
-    console.log('1. Generate a Protractor test:');
-    console.log('yo protractor:unit test-name'.blue);
-    console.log('\n2. Download the Protractor dependencies:');
+    console.log('1. Download the Protractor dependencies:');
     console.log('./node_modules/protractor/bin/install_selenium_standalone'.blue);
-    console.log('\n3. Start the Selenium server:');
+    console.log('\n2. Start the Selenium server:');
     console.log('./selenium/start'.blue);
-    console.log('\n4. Run Protractor:');
+    console.log('\n3. Run Protractor:');
     console.log(('./node_modules/protractor/bin/protractor ' + this.configName).blue);
     console.log('\nYou can read these instructions in README.txt\n');
 
@@ -67,5 +65,6 @@ ProtractorGenerator.prototype.app = function app() {
 ProtractorGenerator.prototype.createFiles = function() {
   this.copy('configTemplate.js', this.configName);
   this.copy('package.json');
+  this.copy('example-spec.js');
   this.copy('README.txt');
 };